backend/requirements.txt
flask
flask-Cors
python-dotenv
PyMuPDF
pdfminer.six
google-genai